address code review
break circular dependency between chat/evaluation by moving ChatInfo to
chat/models.py
